nursing-council-agent/requirements.txt
# Requirements for Nursing Council Agent
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.0
httpx>=0.27.0
pydantic>=2.9.0
azure-storage-blob>=12.19.0
reportlab>=4.0.0
